[refactor]: authMiddleware에서 서비스 로직을 authService로 분리시킴, createToken 유틸함수에서 tokenConfig객체의 프로퍼티명 수정(버그 해결)
- 문제상황: createToken 유틸함수에서 tokenConfig의 프로퍼티 이름과 전달받은 type의 문자열 값이 일치하지 않게 코드를 작성해서 token을 생성할 때 config가 제대로 들어가지 않는 문제 발생
- 해결: 전달받는 type의 문자열 값과 동일한 프로퍼티 이름으로 수정하여 문제 해결
